backend/foodgram/favorites/serializers.py
from rest_framework import serializers
from recipes.models import Recipe
from .models import Favorite

class FavoriteSerializer(serializers.ModelSerializer):
    recipe = serializers.PrimaryKeyRelatedField(queryset=Recipe.objects.all())

    class Meta:
        model = Favorite
        fields = ('user', 'recipe')
